core-net: cast uint32_t grace_us for %u format specifier
connect3.c logs grace_us (uint32_t) with %u. On toolchains where uint32_t
is 'long unsigned int' (e.g. xtensa-esp-elf for ESP-IDF v5.3/v5.4) this
trips -Werror=format and fails the build. Cast to unsigned int to match
the specifier; the value (grace timer in us) always fits.

+++ lib/core-net/client/connect3.c
@@ -1015,7 +1015,7 @@ ads_known:
 					grace_us = 3000000;
 			}
 		}
-		lwsl_wsi_notice(wsi, "QUIC socket created, starting grace timer %uus", grace_us);
+		lwsl_wsi_notice(wsi, "QUIC socket created, starting grace timer %uus", (unsigned int)grace_us);
 		lws_sul_schedule(wsi->a.context, wsi->tsi, &wsi->sul_h3_grace,
 				 lws_client_h3_grace_cb, grace_us);
